+++ kafka/java/crm-camunda/src/main/java/io/flowing/retail/crm/CrmApplication.java
@@ -1,0 +1,14 @@
+package io.flowing.retail.crm;
+
+import org.camunda.bpm.spring.boot.starter.annotation.EnableProcessApplication;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@EnableProcessApplication
+public class CrmApplication {
+
+  public static void main(String[] args) throws Exception {
+    SpringApplication.run(CrmApplication.class, args);
+  }
+}